Factored all MPI code out into batch_job_mpi, with a single excess setup entry point.
